Draw player's gold amount on screen

+++ src/actors/gold.rs
@@ -1,9 +1,9 @@
 use engine::{Actor};
 
 #[derive(Clone)]
-pub struct Gold{ value: u32 }
+pub struct Gold{ value: i32 }
 impl Gold {
-    pub fn new(value: u32) -> Gold {
+    pub fn new(value: i32) -> Gold {
         Gold{ value }
     }
 }
@@ -11,7 +11,7 @@ impl Gold {
 impl Actor for Gold {
     fn can_be_stepped_on(&self) -> bool { true }
     fn be_stepped_on(&mut self, other: &mut Actor) {
-        other.gain_money(self.value as i32);
+        other.gain_money(self.value);
     }
     fn symbol(&self) -> char { 'G' }
 }
